+++ src/Project.js
@@ -1,0 +1,6 @@
+export default class Project {
+  constructor(name,number) {
+    this.name = name;
+    this.number = number;
+  }
+}
